upadate edit

+++ src/views/HomeView.vue
@@ -11,11 +11,13 @@ const getData = async () => {
 };
 onMounted(() => getData());
 const removeData = async (id) => {
-  const response = await axios.delete(
-    `https://api-app.herianto.xyz/api/katalogs/${id}`
-  );
-  console.log(response);
-  getData();
+  if (confirm("Are you sure?")) {
+    const response = await axios.delete(
+      `https://api-app.herianto.xyz/api/katalogs/${id}`
+    );
+    console.log(response);
+    getData();
+  }
 };
 const page = ref(1);
 const perPage = ref(5);
@@ -48,25 +50,25 @@ const changePage = (item) => {
 <template>
   <div class="container">
     <div class="row justify-content-center">
-      <div class="col-md-10">
+      <div class="col-md-12">
         <router-link to="/create" class="btn btn-primary mb-3 mt-3"
           >Create</router-link
         >
         <div class="table-responsive text-center">
           <TableHome :data="paginateData" :removeData="removeData" />
           <div class="text-center">
-            <button @click="prevPage" class="btn btn-outline-warning">
+            <button @click="prevPage" class="btn btn-outline-dark btn-sm me-1">
               &laquo;
             </button>
             <button
               v-for="item in totalPage"
               :key="item"
               @click="changePage(item)"
-              class="btn btn-outline-warning"
+              class="btn btn-outline-dark mx-1 btn-sm"
             >
               {{ item }}
             </button>
-            <button @click="nextPage" class="btn btn-outline-warning">
+            <button @click="nextPage" class="btn btn-outline-dark btn-sm ms-1">
               &raquo;
             </button>
           </div>
